WORKING STATE - Added fixes to random.c/h
Changed Random_String() to take uint32_t arr rather than uint8_t array
Added include for <string.h>
Added temp value for DELAY_LIMIT in RandomDelay()

+++ decoder/inc/random.h
@@ -12,7 +12,7 @@
 
 int RandomInt(void);
 
-void Rand_String(uint8_t *buf, uint32_t len);
+void Rand_String(uint32_t *buf, uint32_t len);
 
 void generate_key(mxc_aes_keys_t keySize, uint32_t address);
